in_tail: Extract group watcher related codes to in_tail/group_watcher.rb
